1.9.2: natural zooming (#71)

+++ OpenGPSLogger/app/src/main/java/eu/tijlb/opengpslogger/ui/view/bitmap/CopyRightNoticeBitmapRenderer.kt
@@ -31,7 +31,7 @@ class CopyRightNoticeBitmapRenderer(val context: Context): AbstractBitmapRendere
 
     override suspend fun draw(
         bbox: BBoxDto,
-        zoom: Int,
+        zoom: Double,
         renderDimension: Pair<Int, Int>,
         assignBitmap: (Bitmap) -> Unit,
         refreshView: () -> Any
